+++ client/wfprev-war/src/main/angular/src/app/components/edit-project/activities/fiscal-activity-item/fiscal-activity-item.component.scss
@@ -179,45 +179,7 @@ input#activityComment {
   border-bottom: 1px solid black; // Underline when editing
 }
 
-.mat-date-range-input {
-  display: flex;
-  justify-content: space-between;
-  font-weight: 400;
-  font-size: 16px;
-  color: #282828;
-  flex: 1 1 auto !important;
-}
 
-.mat-date-range-input input {
-  border: none !important; /* Remove borders */
-  box-shadow: none !important;
-  text-align: center;
-  background: transparent;
-  margin: 0 !important;
-}
-
-.mat-form-field-appearance-outline .mat-form-field-outline {
-  display: none !important; /* Hide the default outlined box */
-}
-
-.mat-datepicker-toggle {
-  margin-left: 8px; /* Add space between date input and calendar icon */
-}
-
-.date-range-container {
-  display: flex;
-  align-items: center;
-  width: 100%;
-}
-
-.date-range-container mat-form-field {
-  flex: 1 1 auto;
-  width: 100%;
-}
-
-.date-separator {
-  margin: 0 5px;
-}
 
 input[readonly] {
   border-bottom: 1px dashed #2d2d2d !important;
@@ -312,32 +274,7 @@ input[readonly] {
   }
 }
 
-.custom-overlay-container .cdk-overlay-pane {
-  left: 190px !important;
-  transform: none !important;
-}
 
-:host ::ng-deep .mdc-text-field--filled:not(.mdc-text-field--disabled) {
-  background-color: white;
-  padding: 8px 12px;
-}
-
-:host ::ng-deep .mat-mdc-form-field-flex {
-  height: 20px;
-  margin: 4px;
-}
-:host ::ng-deep .mat-mdc-form-field-infix {
-  padding: 0 !important;
-}
-
-.date-range-container .mat-date-range-input input {
-  padding: 0 !important; /* remove inner padding */
-}
-
-::ng-deep .mat-form-field-disabled .mdc-text-field {
-  margin-top: 6px;
-  height: 38px;
-}
 
 .wf-empty-state {
   margin-top: 0px;
