+++ client/wfprev-war/src/main/angular/src/app/components/year-end-performance-update/year-end-activity-item/year-end-activity-item.component.scss
@@ -254,65 +254,5 @@
   margin-bottom: 15px;
 }
 
-.mat-date-range-input {
-  display: flex;
-  justify-content: space-between;
-  font-weight: 400;
-  font-size: 16px;
-  color: #282828;
-  flex: 1 1 auto !important;
-}
 
-.mat-date-range-input input {
-  border: none !important; /* Remove borders */
-  box-shadow: none !important;
-  text-align: left;
-  background: transparent;
-  margin: 0 !important;
-}
-
-.mat-form-field-appearance-outline .mat-form-field-outline {
-  display: none !important; /* Hide the default outlined box */
-}
-
-.mat-datepicker-toggle {
-  margin-left: 8px; /* Add space between date input and calendar icon */
-}
-
-.date-range-container {
-  display: flex;
-  align-items: center;
-  width: 100%;
-}
-
-.date-range-container mat-form-field {
-  flex: 1 1 auto;
-  width: 100%;
-}
-
-.date-separator {
-  margin: 0 5px;
-}
-
-:host ::ng-deep .mdc-text-field--filled:not(.mdc-text-field--disabled) {
-  background-color: white;
-  padding: 8px 12px;
-}
-
-:host ::ng-deep .mat-mdc-form-field-flex {
-  height: 20px;
-  margin: 4px;
-}
-:host ::ng-deep .mat-mdc-form-field-infix {
-  padding: 0 !important;
-}
-
-.date-range-container .mat-date-range-input input {
-  padding: 0 !important; /* remove inner padding */
-}
-
-::ng-deep .mat-form-field-disabled .mdc-text-field {
-  margin-top: 6px;
-  height: 38px;
-}
